http://www.jwz.org/xscreensaver/xscreensaver-5.09.tar.gz -rw-r--r-- 1 zblaxell zblaxell 5549665 Sep  3 05:08 xscreensaver-5.09.tar.gz 24cb3d04244ee6f89c16866f0eed93a8d16d0de2  xscreensaver-5.09.tar.gz

+++ OSX/InvertedSlider.m
@@ -70,7 +70,7 @@
 /* Implment all setter methods in terms of "setDoubleValue", above.
  */
 
--(void) setFloatValue:(double)v
+-(void) setFloatValue:(float)v
 {
   [self setDoubleValue:(double)v];
 }
